launcher: delete config provider, move logic to binary provider
