+++ clients/launcher/test/test_utils.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
-import 'package:launcher/providers/config_provider.dart';
 import 'package:launcher/providers/quotes_provider.dart';
 import 'package:logger/logger.dart';
 import 'package:provider/provider.dart';
 import 'package:sail_ui/providers/binary_provider.dart';
 import 'package:sail_ui/sail_ui.dart';
 
-import 'mocks/configuration_service_mock.dart';
 import 'mocks/download_manager_mock.dart';
 import 'mocks/quotes_provider_mock.dart';
 import 'mocks/storage_mock.dart';
@@ -53,12 +51,6 @@ Future<void> registerTestDependencies() async {
   if (!GetIt.I.isRegistered<ClientSettings>()) {
     GetIt.I.registerLazySingleton<ClientSettings>(
       () => ClientSettings(store: MockStore(), log: log),
-    );
-  }
-
-  if (!GetIt.I.isRegistered<ConfigProvider>()) {
-    GetIt.I.registerLazySingleton<ConfigProvider>(
-      () => MockConfigProvider(),
     );
   }
